revise migrations,models,controllers

+++ app/database/migrations/2025_09_22_033901_create_order_details_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('order_id')->constrained('orders')->cascadeOnDelete();
             $table->foreignId('service_id')->constrained('services')->cascadeOnDelete();
             $table->integer('quantity')->default(1);
-            $table->decimal('subtotal', 10, 2);
+            $table->float('subtotal');
             $table->timestamps();
         });
     }
